spotify/requirements.txt
Django~=3.2.7
djangorestframework~=3.12.4
drf-yasg~=1.20.0
psycopg2-binary==2.9.1
